Doing a clean if the environment is out-of-sync
There've been a lot of errors recently where updates cause the cached
pickle and makefiles to become unusable. This is a 'panic' mechanism
to reset the .build and start fresh.

Still TODO is to do a similar reset if and when the environment source
changes (i.e. Arduino IDE version is different).